Adding Worker as an option in the SDK
This is an intermediate step towards replacing 'slave' with 'worker' in the cluster manager types. It will allow us to update the CLI and to add workerCount and related commands. After the CLI switches to 'worker',and ruby integration test change from slave to worker. 'slave' will be removed permanently.

Change-Id: I96f3279f8805d3333f10f72adff986eb5aa192a5