Remove SlaveCount and SlaveBatchSize from the SDK
Remove the Slavecount and slaveBatchSize from the SDK.
I also changed it so that  workerCount becomes a required field of in the createClusterSpec
And NewWorkerCount becomes a required field of the resizeoperation.
This, however, doesn't match the APIFE. In APIFE we still set those fields as optional.
This shouldnt cause anything to break. I will change APIFE next.

Change-Id: If515b7149a47b4c88f59b2bbc1a4ddabecbf6208