Allow tokens to be refreshed
This enables refreshing of tokens. Lightwave grants two tokens--an
access token (short-lived, used to access the API) and a refresh token
(longer-lived, allows you to get a new access token.

Conceptually this is straightforward: on every REST API call to Photon
Controller, if we get back a 401 status code and the Photon Controller
tells us "ExpiredAuthToken", just get a new token and retry the
call. That is it, but there are a few more details:

* Throughout the SDK we passed just the access token to the REST
  client, and we need to pay both the acces and refresh tokens.
* All the SDK methods expect to be able to read (in the sense of a
  Golan Reader interface) the response, but we have to read the error,
  and the HTTP reader doesn't allow seek. When we get the error back,
  we create a new Reader to allow this.
* When we need to retry a POST, we have to "rewind" (seek to the
  beginning) of the POST body, so the HTTP client can re-send it. It's
  a bit awkward when sending images, so there's some complexity in the
  REST client to handle it. This required changing some uses of
  Buffers to Readers--Buffers were not needed, so this was a
  transparent change.
* When we refresh the token, we need to inform the user of the
  SDK. Our main SDK user is the command-line interface. It will stash
  the new token in it's configuration file, so future invocations of
  the CLI will use the new token.

Testing this is tricky within our unit tests. I've thoroughly tested
this by hand with the CLI. I think the right test is to modify our
integration tests to no longer have access tokens that are
artificially long-lived--that will come after everything else is
committed.

Change-Id: I1b25df911144b6993a2b7249b4246d070f58a071